% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP.plot.R
\name{CP.plot}
\alias{CP.plot}
\title{A Change Point (CP) Function}
\usage{
CP.plot(data, peptide, L = 1, U = 5, metric, normalization = TRUE,
  ytitle = "Change Point Plot", type = "mean")
}
\arguments{
\item{data}{Comma-separated (*.csv), QC file format. It should contain a Precursor column and the metrics columns.}

\item{peptide}{The name of precursor you want to draw the plot for}

\item{L}{Lower bound of the giude set}

\item{U}{Upper bound of the guide set}

\item{ytitle}{The y-axis title of the plot. The x-axis title is by default "QCno-name of peptide"}

\item{type}{can take two values, "mean" or "dispersion".}
}
\description{
This function allows you to calculate change point statistics for a step change.
}
\examples{
CP.plot()
}
\keyword{Change}
\keyword{chart}
\keyword{control}
\keyword{point}

