% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPTAC.Study9.1.Site54.R
\docType{data}
\name{CPTAC.Study9.1.Site54}
\alias{CPTAC.Study9.1.Site54}
\title{Arabidopsis QTL data on gravitropism}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{http://qtlarchive.org/db/q?pg=projdetails&proj=moore_2013b}{QTL Archive}
}
\usage{
data(CPTAC.Study9.1.Site54)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(CPTAC.Study9.1.Site54)
times <- attr(CPTAC.Study9.1.Site54, "precursor")
pre <- grav$precursor
\donttest{iplotCurves(phe, times)}
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{http://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
