% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlots.R
\name{CUSUM.Summary.plot}
\alias{CUSUM.Summary.plot}
\title{CUSUM summary plot}
\usage{
CUSUM.Summary.plot(data, L, U)
}
\arguments{
\item{data}{Comma-separated (*.csv), QC file format. It should contain a Precursor column and the metrics columns.}

\item{L}{Lower bound of the giude set}

\item{U}{Upper bound of the guide set}
}
\description{
This function allows you to draw the CUSUM summary plot.
}
\examples{
CUSUM.Summary.plot.plot()
}
\keyword{CUSUM}
\keyword{Summary}
\keyword{plot}

