% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CUSUMRiverPlot.R
\name{CUSUMRiverPlot}
\alias{CUSUMRiverPlot}
\title{A function to create river plot to aggregate results from CUSUM charts}
\usage{
CUSUMRiverPlot(data = NULL, L = 1, U = 5, listMean = NULL,
  listSD = NULL)
}
\arguments{
\item{data}{omma-separated (.csv), metric file. It should contain a "Precursor" column and the metrics columns. It should also include "Annotations" for each observation.}

\item{L}{lower bound of the guide set.}

\item{U}{upper bound of the guide set.}

\item{listMean}{list of the means for each metric. It is used when mean is known. It is NULL when mean is not known.  The default is NULL.}

\item{listSD}{list of the standard deviations for each metric. It is used when standard deviation is known. It is NULL when mean is not known. The default is NULL.}
}
\description{
A function to create river plot to aggregate results from CUSUM charts
}
\examples{
# First process the data to make sure it's ready to use
sampleData <- DataProcess(S9Site54)
head(sampleData)
# Draw CUSUM summary plot
CUSUMRiverPlot(data = sampleData)
CUSUMRiverPlot(data = sampleData,
                listMean = list("BestRetentionTime" = 27.78,
                                "TotalArea" = 35097129,
                                "MaxFWHM" = 0.28,
                                "Peak Assymetry" = 0.98),
                listSD = list("BestRetentionTime" = 8.19,
                              "TotalArea" = 34132861,
                              "MaxFWHM" = 0.054,
                              "Peak Assymetry" = 0.002)
                )
}
\keyword{CUSUM}
