% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlots.R
\name{CUSUMRiverPlots}
\alias{CUSUMRiverPlots}
\title{A river plot to summarize results from CUSUMm and CUSUMv plots}
\usage{
CUSUMRiverPlots(data = NULL, L = 1, U = 5, listMean = NULL,
  listSD = NULL)
}
\arguments{
\item{data}{Comma-separated (*.csv), QC file format.
It should contain a Precursor column and the metrics columns.}

\item{L}{Lower bound of the giude set. Default is 1.}

\item{U}{Upper bound of the guide set. Default is 5.}

\item{listMean}{List of the means for the metrics.
If you don't know the means leave it as NULL and they will be calculated automatically by using L and U. The default is NULL.}

\item{listSD}{List of the standard deviations for the metrics.
If you don't know the standard deviations leave it as NULL and they will be calculated automatically by using L and U. The default is NULL.}
}
\description{
This function allows you to draw the CUSUM summary plot.
}
\keyword{CUSUM}
\keyword{Summary}
\keyword{plot}

