% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.plot.R
\name{DecisionMaker}
\alias{DecisionMaker}
\title{Decision maker}
\usage{
DecisionMaker(data, method, peptideThresholdRed = 0.7,
  peptideThresholdYellow = 0.5, L = 1, U = 5, type = "mean",
  title = "heatmap plot", selectMean = NULL, selectSD = NULL)
}
\arguments{
\item{data}{Comma-separated (*.csv), QC file format. It should contain a Precursor column and the metrics columns.}

\item{method}{It is either "CUSUM" or "XmR"}

\item{peptideThresholdRed}{Is a threshold that marks percentage of peptides above it red on the heatmap. Defaults to 0.7}

\item{peptideThresholdYellow}{Is a threshold that marks percentage of peptides above it and below the peptideThresholdRed, yellow on the heatmap. Defaults to 0.5}

\item{L}{Lower bound of the giude set. Defaults to 1}

\item{U}{Upper bound of the guide set. Defaults to 5}

\item{type}{can take two values, "mean" or "dispersion". Defaults to "mean"}

\item{title}{the title of the plot. Defaults to "heatmap plot"}

\item{selectMean}{If you want to choose mean yourself and do not use the guide set. Defaults to NULL}

\item{selectSD}{If you want to choose standard deviation yourself and do not want to use the guide set. Defaults to NULL}
}
\description{
This function allows you to draw the heatmap plot.
}
\examples{
DecisionMaker()
}
\keyword{heatmap}
