% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingDataMap.R
\name{MissingDataMap}
\alias{MissingDataMap}
\title{A function to summarize missing values}
\usage{
MissingDataMap(data)
}
\arguments{
\item{data}{Processed data}
}
\value{
A plot of missing values.
}
\description{
A function to summarize missing values
}
\examples{
# The data is "S9Site54" which is defined in the package.
#data <- DataProcess(S9Site54)
#MissingDataMap(data)
}
