% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSnbaseToMSstatsQC.R
\name{MSnbaseToMSstatsQC}
\alias{MSnbaseToMSstatsQC}
\title{A function to convert MSnbase files to MSstatsQC format}
\usage{
MSnbaseToMSstatsQC(msfile, identfile)
}
\arguments{
\item{msfile}{data file to be converted}

\item{identfile}{data file for features}
}
\value{
A data frame that can be used with MSstatsQC

A csv file that is converted from raw files
}
\description{
A function to convert MSnbase files to MSstatsQC format
}
\examples{
library("RforProteomics")
msfile <- getPXD000001mzXML()
MSnbaseToMSstatsQC(msfile)
}
\keyword{MSnbase,}
\keyword{input}
\keyword{qcmetrics,}
