% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstatsQC.ML.deployR.R
\name{MSstatsQC.ML.deployR}
\alias{MSstatsQC.ML.deployR}
\title{A function to test random forest classifiers for QC data}
\usage{
MSstatsQC.ML.deployR(Test.set, guide.set, rf_model)
}
\arguments{
\item{Test.set}{comma-separated (.csv), metric file. It should contain a "Precursor" column and the metrics columns. It should also include "Annotations" for each run.}

\item{guide.set}{comma-separated (.csv), metric file. It should contain a "Precursor" column and the metrics columns. It should also include "Annotations" for each run.}

\item{rf_model}{the model that was trained previosly by MSstatsQC-ML training process}
}
\value{
Probability of failure predictions based on a trained model
}
\description{
A function to test random forest classifiers for QC data
}
\examples{
S9Site54.dataML <- DataProcess(MSstatsQC::S9Site54[, ])
colnames(S9Site54.dataML)[1] <- c("idfile")
colnames(S9Site54.dataML)[2] <- c("peptide")
S9Site54.dataML$peptide <- as.factor(S9Site54.dataML$peptide)
S9Site54.dataML$idfile <- as.numeric(S9Site54.dataML$idfile)
S9Site54.dataML <- within(S9Site54.dataML, rm(Annotations, missing))
guide.set <- dplyr::filter(S9Site54.dataML, idfile <= 20)
\donttest{
rf_model <- MSstatsQC.ML.trainR(guide.set, sim.size = 10)
}
Test.set <- dplyr::filter(S9Site54.dataML, idfile > 20)
\donttest{
MSstatsQC.ML.deployR(Test.set, guide.set, rf_model = rf_model)
}
Test.set <- S9Site54.dataML
\donttest{
MSstatsQC.ML.deployR(Test.set, guide.set, rf_model = rf_model)
}

}
\keyword{deployment}
\keyword{learning,}
\keyword{machine}
