% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstatsQC.ML.sim.size.detectR.R
\name{MSstatsQC.ML.sim.size.detectR}
\alias{MSstatsQC.ML.sim.size.detectR}
\title{A function to train random forest classifiers for QC data}
\usage{
MSstatsQC.ML.sim.size.detectR(guide.set, sim.start, sim.end)
}
\arguments{
\item{guide.set}{comma-separated (.csv), metric file. It should contain a "Precursor" column and the metrics columns. It should also include "Annotations" for each run.}

\item{sim.start}{enter min simulation size.}

\item{sim.end}{enter max simulation size.}
}
\value{
a plot for sim.size vs performance
}
\description{
A function to train random forest classifiers for QC data
}
\examples{
# First process the data to make sure it's ready to use
S9Site54.dataML <- DataProcess(MSstatsQC::S9Site54[, ])
colnames(S9Site54.dataML)[1] <- c("idfile")
colnames(S9Site54.dataML)[2] <- c("peptide")
S9Site54.dataML$peptide <- as.factor(S9Site54.dataML$peptide)
S9Site54.dataML$idfile <- as.numeric(S9Site54.dataML$idfile)
S9Site54.dataML <- within(S9Site54.dataML, rm(Annotations, missing))
guide.set <- dplyr::filter(S9Site54.dataML, idfile <= 20)
\donttest{
MSstatsQC.ML.sim.size.detectR(guide.set, sim.start = 10, sim.end = 2500)
}
}
\keyword{learning,}
\keyword{machine}
\keyword{simulation}
\keyword{size}
