% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstatsQC.ML.trainR.R
\name{MSstatsQC.ML.trainR}
\alias{MSstatsQC.ML.trainR}
\title{A function to train random forest classifiers for QC data}
\usage{
MSstatsQC.ML.trainR(
  guide.set,
  sim.size,
  guide.set.annotations = NULL,
  nfolds = NULL,
  a = 1.5,
  b = 2
)
}
\arguments{
\item{guide.set}{comma-separated (.csv), metric file. It should contain a "Precursor" column and the metrics columns. It should also include "Annotations" for each run.}

\item{sim.size}{enter simulation size.}

\item{guide.set.annotations}{comma-separated (.csv), metric file with annotations such as pass and fail.}

\item{nfolds}{fold for cross validation}

\item{a}{lower threshold to define shift size}

\item{b}{upper threshold to define shift size}
}
\value{
A trained model and performance indicators from train/validation/test splits
}
\description{
A function to train random forest classifiers for QC data
}
\examples{
S9Site54.dataML <- DataProcess(MSstatsQC::S9Site54[, ])
colnames(S9Site54.dataML)[1] <- c("idfile")
colnames(S9Site54.dataML)[2] <- c("peptide")
S9Site54.dataML$peptide <- as.factor(S9Site54.dataML$peptide)
S9Site54.dataML$idfile <- as.numeric(S9Site54.dataML$idfile)
S9Site54.dataML <- within(S9Site54.dataML, rm(Annotations, missing))
guide.set <- dplyr::filter(S9Site54.dataML, idfile <= 20)
\donttest{
MSstatsQC.ML.trainR(guide.set, sim.size = 10)
}

}
\keyword{engineering}
\keyword{feature}
\keyword{learning,}
\keyword{machine}
\keyword{training,}
