% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQCToMSstatsQC.R
\name{mzQCToMSstatsQC}
\alias{mzQCToMSstatsQC}
\title{A function to convert mzQC files to MSstatsQC format}
\usage{
mzQCToMSstatsQC(mzQCfile)
}
\arguments{
\item{mzQCfile}{data file to be converted}
}
\value{
A data frame that can be used with MSstatsQC

A csv file that is converted from raw files
}
\description{
A function to convert mzQC files to MSstatsQC format
}
\examples{
\donttest{
library(RforProteomics)
}
\donttest{
msfile <- getPXD000001mzXML()
}
\donttest{
mzQCToMSstatsQC(msfile)
}
}
\keyword{input}
\keyword{mzqc,}
\keyword{qcmetrics,}
