% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCloudDDA.R
\docType{data}
\name{QCloudDDA}
\alias{QCloudDDA}
\title{DDA QC data from QCloud System}
\format{csv}
\usage{
data(QCloudDDA)
}
\value{
An example dataset generated from QCloud system
}
\description{
QC results generated from QCloud system
}
\details{
DDA QC data from QCloud System
}
\keyword{datasets}
