% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCloudSRM.R
\docType{data}
\name{QCloudSRM}
\alias{QCloudSRM}
\title{SRM QC data from QCloud System}
\format{
csv
}
\usage{
data(QCloudSRM)
}
\value{
An example dataset generated from QCloud system
}
\description{
QC results generated from QCloud system
}
\details{
SRM QC data from QCloud System
}
\examples{
head(QCloudSRM)
}
\keyword{datasets}
