% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuiCDIA.R
\docType{data}
\name{QuiCDIA}
\alias{QuiCDIA}
\title{DIA iRT data from QuiC System}
\format{
csv
}
\usage{
data(QuiCDIA)
}
\value{
An example dataset generated from QuiC system
}
\description{
QC results generated from QuiC system
}
\details{
DIA iRT data from QuiC System
}
\keyword{datasets}
