% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveMissing.R
\name{RemoveMissing}
\alias{RemoveMissing}
\title{A data processing function for removing missing values}
\usage{
RemoveMissing(data = NULL)
}
\arguments{
\item{data}{Comma-separated (*.csv), QC file format. It should contain a Precursor column and the metrics columns.}
}
\value{
A data frame that processes using \code{input.sanity.check} function.
}
\description{
A data processing function for removing missing values
}
\examples{
# The data is "S9Site54" which is defined in the package.
data <- RemoveMissing(S9Site54)
}
