% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunShiny.R
\name{RunShiny}
\alias{RunShiny}
\title{A function to launch MSstatsQC Shiny app.}
\usage{
RunShiny(data = NULL)
}
\arguments{
\item{data}{Comma-separated (*.csv), QC file format. It should contain a Precursor column
and the metrics columns.}
}
\description{
A function to launch MSstatsQC Shiny app.
}
\keyword{shiny}
