% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S9Site54.R
\docType{data}
\name{S9Site54}
\alias{S9Site54}
\title{CPTAC study 9.1 site 54 dataset}
\format{csv}
\usage{
data(S9Site54)
}
\description{
system suitabilty testing results generated during CPTAC Study 9.1 for Site 54
}
\details{
CPTAC system suitability testing data for Site 54 from Study 9.1
}
\references{
\url{http://www.mcponline.org/content/early/2015/02/18/mcp.M114.047050}
}
\keyword{datasets}
