% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XmRChart.R
\name{XmRChart}
\alias{XmRChart}
\title{A function to construct individual (X) and moving range (mR) control charts}
\usage{
XmRChart(data = NULL, peptide, L = 1, U = 5, metric,
  normalization = FALSE, ytitle = "Individual observations",
  type = "mean", selectMean = NULL, selectSD = NULL)
}
\arguments{
\item{data}{comma-separated (.csv), metric file. It should contain a "Precursor" column and the metrics columns. It should also include "Annotations" for each observation.}

\item{peptide}{the name of precursor of interest.}

\item{L}{Lower bound of the guide set.}

\item{U}{Upper bound of the guide set.}

\item{metric}{the name of metric of interest.}

\item{normalization}{TRUE if metric is standardized and FALSE if not standardized.}

\item{ytitle}{the y-axis title of the plot.  Defaults to "Individual observations". The x-axis title is by default "Time : name of peptide"}

\item{type}{the type of the control chart. Two values can be assigned, "mean" or "variability". Default is "mean".}

\item{selectMean}{the mean of a metric. It is used when mean is known. It is NULL when mean is not known.  The default is NULL.}

\item{selectSD}{the standard deviation of a metric. It is used when standard deviation is known. It is NULL when mean is not known. The default is NULL.}
}
\value{
A plot of individual values or moving ranges versus time per peptide and metric generated from \code{XmR.data.prepare} data frame.
}
\description{
A function to construct individual (X) and moving range (mR) control charts
}
\examples{
# First process the data to make sure it's ready to use
sampleData <- DataProcess(S9Site54)
head(sampleData)
# Find the name of the peptides
levels(sampleData$Precursor)
# Calculate X and mR statistics
XmRChart(data = sampleData, peptide = "VLVLDTDYK", metric = "BestRetentionTime")
XmRChart(data = sampleData, peptide = "VLVLDTDYK", metric = "BestRetentionTime",
         ytitle = "moving ranges", type = "variability")
XmRChart(data = sampleData, peptide = "VLVLDTDYK", metric = "BestRetentionTime",
         selectMean = 27.78, selectSD = 8.19)
XmRChart(data = sampleData, peptide = "DDGSWEVIEGYR", metric = "TotalArea")
XmRChart(data = sampleData, peptide = "DDGSWEVIEGYR", metric = "TotalArea",
         selectMean = 35097129, selectSD = 34132861)
XmRChart(data = sampleData, peptide = "TAAYVNAIEK", metric = "MaxFWHM")
XmRChart(data = sampleData, peptide = "LVNELTEFAK", metric = "Peak Assymetry")
}
\keyword{XmR,}
\keyword{chart}
\keyword{control}
