% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePointEstimator.R
\name{ChangePointEstimator}
\alias{ChangePointEstimator}
\title{A Change Point (CP) Function}
\usage{
ChangePointEstimator(data, peptide, L = 1, U = 5, metric,
  normalization = TRUE, ytitle = "Change Point Plot - mean",
  type = "mean")
}
\arguments{
\item{data}{Comma-separated (*.csv), QC file format. It should contain a Precursor column and the metrics columns.}

\item{peptide}{The name of precursor you want to draw the plot for}

\item{L}{Lower bound of the guide set}

\item{U}{Upper bound of the guide set}

\item{metric}{QC metric to be monitored}

\item{normalization}{TRUE if QC metric is standardized and FALSE if not standardized}

\item{ytitle}{The y-axis title of the plot. Defaults to "Change Point Plot - mean". The x-axis title is by default "QCno-name of peptide"}

\item{type}{can take two values, "mean" or "dispersion".}
}
\description{
This function allows you to calculate change point statistics for a step change in mean or variability.
}
\examples{
ChangePointEstimator()
}
\keyword{Change}
\keyword{chart}
\keyword{control}
\keyword{point,}

