% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CUSUMPlots.R
\name{CUSUMPlots}
\alias{CUSUMPlots}
\title{A Cumulative Sum (CUSUM) Function}
\usage{
CUSUMPlots(data, peptide, L = 1, U = 5, metric, normalization = TRUE,
  ytitle = "CUSUM Plot - mean", type = "mean")
}
\arguments{
\item{data}{Comma-separated (*.csv), QC file format. It should contain a Precursor column and the metrics columns.}

\item{peptide}{The name of precursor you want to draw the plot for}

\item{L}{Lower bound of the giude set. Defaults to L = 1}

\item{U}{Upper bound of the guide set. Defaults to U = 5}

\item{ytitle}{The y-axis title of the plot.Defaults to "CUSUM Plot - mean". The x-axis title is by default "QCno-name of peptide"}

\item{type}{can take two values, "mean" or "dispersion". Defaults to "mean"}
}
\description{
This function allows you to calculate CUSUM statistics.
}
\examples{
CUSUMPlots()
}
\keyword{Cumulative}
\keyword{Sum,}
\keyword{chart}
\keyword{control}
