% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlots.R
\name{CUSUMSummaryPlots}
\alias{CUSUMSummaryPlots}
\title{A river plot to summarize results from CUSUMm and CUSUMv plots}
\usage{
CUSUMSummaryPlots(data = NULL, L = 1, U = 5, listMean = NULL,
  listSD = NULL)
}
\arguments{
\item{data}{Comma-separated (*.csv), QC file format.
It should contain a Precursor column and the metrics columns.}

\item{L}{Lower bound of the giude set. Default is 1.}

\item{U}{Upper bound of the guide set. Default is 5.}

\item{listMean}{List of the means for the metrics.
If you don't know the means leave it as NULL and they will be calculated automatically by using L and U. The default is NULL.}

\item{listSD}{List of the standard deviations for the metrics.
If you don't know the standard deviations leave it as NULL and they will be calculated automatically by using L and U. The default is NULL.}
}
\description{
This function allows you to draw the CUSUM summary plot.
}
\examples{
# First process the data to make sure it's ready to use
sampleData <- DataProcess(S9Site54)
head(sampleData)
# Draw CUSUM summary plot
CUSUMSummaryPlots(data = sampleData)
CUSUMSummaryPlots(data = sampleData,
                listMean = list("BestRetentionTime" = 27.78,
                                "TotalArea" = 35097129,
                                "MaxFWHM" = 0.28,
                                "Peak Assymetry" = 0.98),
                listSD = list("BestRetentionTime" = 8.19,
                              "TotalArea" = 34132861,
                              "MaxFWHM" = 0.054,
                              "Peak Assymetry" = 0.002)
                )
}
\keyword{CUSUM}
\keyword{Summary}
\keyword{plot}

