% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlots.R
\name{CUSUMSummaryPlots}
\alias{CUSUMSummaryPlots}
\title{A river and a radar plot to summarize results from CUSUMm and CUSUMv plots}
\usage{
CUSUMSummaryPlots(data, L = 1, U = 5)
}
\arguments{
\item{data}{Comma-separated (*.csv), QC file format. It should contain a Precursor column and the metrics columns.}

\item{L}{Lower bound of the giude set. Default is 1.}

\item{U}{Upper bound of the guide set. Default is 5.}
}
\description{
This function allows you to draw the CUSUM summary plot.
}
\examples{
CUSUMSummaryPlots()
}
\keyword{CUSUM}
\keyword{Summary}
\keyword{plot}

