% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{DataProcess}
\alias{DataProcess}
\title{A data processing function}
\usage{
DataProcess(data)
}
\arguments{
\item{data}{Comma-separated (*.csv), QC file format. It should contain a Precursor column and the metrics columns.}
}
\description{
This function allows you to validate the dataset for further analysis
}
\examples{
DataProcess()
}

