% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSnbaseToMSstatsQC.R
\name{MSnbaseToMSstatsQC}
\alias{MSnbaseToMSstatsQC}
\title{A function to convert MSnbase files to MSstatsQC format}
\usage{
MSnbaseToMSstatsQC(msfile)
}
\arguments{
\item{msfile}{data file to be converted}
}
\value{
A data frame that can be used with MSstatsQC
}
\description{
A function to convert MSnbase files to MSstatsQC format
}
\examples{
dontrun{MSstatsQCdata<- MSnbaseToMSstatsQC(msfile)}
}
\keyword{MSnbase,}
\keyword{input}
\keyword{qcmetrics,}
