% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunShiny.R
\name{RunShiny}
\alias{RunShiny}
\title{Shiny connection for example datasets}
\usage{
RunShiny(data = NULL)
}
\arguments{
\item{data}{Comma-separated (*.csv), QC file format. It should contain a Precursor column
and the metrics columns.}
}
\description{
This function allows you to draw the heatmaps to help user test their decision intervals.
 This plot provides massages about overall system performance.
}
\keyword{shiny}

