% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S9Site54.R
\docType{data}
\name{S9Site54}
\alias{S9Site54}
\title{CPTAC system suitability testing data for Site 54 from Study 9.1}
\format{csv}
\source{
from labratory
}
\usage{
data(S9Site54)
}
\description{
a sample data to use in the package
}
\details{
CPTAC system suitability testing data for Site 54 from Study 9.1
}
\references{
my computer
}
\keyword{datasets}

