% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlots.R
\name{XmRSummaryPlots}
\alias{XmRSummaryPlots}
\title{XmR summary plot}
\usage{
XmRSummaryPlots(data, L, U)
}
\arguments{
\item{data}{Comma-separated (*.csv), QC file format. It should contain a Precursor column and the metrics columns.}

\item{L}{Lower bound of the giude set}

\item{U}{Upper bound of the guide set}
}
\description{
This function allows you to draw the XmR summary plot.
}
\examples{
XmRSummaryPlots()
}
\keyword{Summary}
\keyword{XmR}
\keyword{plot}
